// utils/types/enums/buttons.enum.ts
export enum ButtonVariantEnum {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  TERTIARY = 'tertiary',
  CERULEAN = 'cerulean',
  GREEN = 'green',
}

export enum ButtonSizeEnum {
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
  XLARGE = 'xlarge',
  XXLARGE = 'xxlarge'
}

export enum ButtonTagEnum {
  ANCHOR = 'a',
  BUTTON = 'button',
  NUXT_LINK = 'nuxt-link',
}

export enum ButtonSaveTypeEnum {
  ADVISER = 'adviser',
  BLOG = 'blog',
  PRODUCT = 'product',
}